+++ spec/associations/server_auth/ratings_spec.rb
@@ -1,2 +1,0 @@
-# Ratings cannot be tested with a server app because
-# only authenticated clients can like and unlike videos